src/SWAT_auth/server/requirements.txt
pymongo
numpy
pillow
tensorflow
